+++ src/services/readLocalFileService.js
@@ -1,0 +1,9 @@
+const helper = require('./../helpers/FileHelper')
+
+class ReadLocalFileService {
+    async getUserInfo() {
+        return helper.get()
+    }
+}
+
+module.exports = new ReadLocalFileService()
